Add ability to copy values from parent endpoint response to sub-endpoint response
The main use case for this is passing parent object properties (i.e. project name)
to sub-endpoints (like project statistics) which are missing parent object props
in order to be uniquely identifiable by the customer. For an example of such a case,
check Gitlab's Project statistics API call in v4